requirements.txt
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
lmdb>=1.2.0
PyYAML>=5.4.0
tqdm>=4.62.0 